Place tests in a new files Add TestCases and Categories in NumberValidatorTests Add static Create instead of constructor
